requirements.txt
django==3.2.18
python-dotenv==0.21.1
psycopg2==2.9.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
social-auth-app-django==5.2.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
click==8.1.3
gspread==5.8.0
google_api_python_client==2.85.0
gspread_formatting==1.1.2